Added unbind() function. Added with() to execute a thunk.
unbind() undoes the effect of exactly one rebind().

rebinds() can be stacked. unbind() undoes the last rebind().

While a rebinding is active, calling console.log() or console.info()
will result in process.stdout() temporarily being replaced with stderr()
for the duration of a call to the original console.log() implementation.


Signed-off-by: Jon Seymour <jon.seymour@gmail.com>